requirements.txt
librouteros
django
djangorestframework
djangorestframework-simplejwt
pillow
reportlab
dateutil
django-cors-headers
mysqlclient
gunicorn
